edited unit test analyzer and added more documentation

+++ toolchain/unitTestAnalyzer/unitTestAnalyzer.bash
@@ -4,11 +4,49 @@ if [ -z "$1" ]; then
     exit 1
 fi
 
-dir="$1"
-pathToUnitTestOverheadInserter="../unitTestOverheadInserter/unitTestOverheadInserter.go"
-pathToUnitTestOverheadRemover="../unitTestOverheadRemover/unitTestOverheadRemover.go"
-pathToPatchedGoRuntime="../../go-patch/bin/go"
-pathToGoRoot="../../go-patch/"
+
+while [[ $# -gt 0 ]]; do
+  key="$1"
+  case $key in
+    -p|--patched-go-runtime)
+      pathToPatchedGoRuntime="$2"
+      shift
+      shift
+      ;;
+    -g|--go-root)
+      pathToGoRoot="$2"
+      shift
+      shift
+      ;;
+    -i|--overhead-inserter)
+      pathToOverHeaderInserter="$2"
+      shift
+      shift
+      ;;
+    -r|--overhead-remover)
+      pathToOverheadRemover="$2"
+      shift
+      shift
+      ;;
+    -f|--folder)
+      dir="$2"
+      shift
+      shift
+      ;;
+    *)
+      shift
+      ;;
+  esac
+done
+
+if [ -z "$pathToPatchedGoRuntime" ] || [ -z "$pathToGoRoot" ] || [ -z "$pathToOverHeaderInserter" ] || [ -z "$pathToOverheadRemover" ] || [-z "$dir" ];then
+  echo "Usage: $0 -p <patched-go-runtime> -g <go-root> -i <overhead-inserter> -r <overhead-remover> -f <folder>"
+  exit 1
+fi
+
+
+
+
 cd "$dir"
 #make folder to store the result
 rm -r "advocateResults"
@@ -21,10 +59,10 @@ for file in $test_files; do
     test_functions=$(grep -oE "[a-zA-Z0-9_]+ *Test[a-zA-Z0-9_]*" $file | sed 's/ *\(t *\*testing\.T\)//' | sed 's/func //')
     for test_func in $test_functions; do
         echo "Running test: $test_func" in "$file of package $package_path"
-        $pathToPatchedGoRuntime run "$pathToUnitTestOverheadInserter" -f "$file" -t "$test_func"
+        $pathToPatchedGoRuntime run "$pathToOverheadInserter" -f "$file" -t "$test_func"
         echo "go test -count=1 -run=$test_func $package_path"
         $pathToPatchedGoRuntime test -count=1 -run="$test_func" "$package_path"
-        $pathToPatchedGoRuntime run "$pathToUnitTestOverheadRemover" -f "$file" -t "$test_func"
+        $pathToPatchedGoRuntime run "$pathToOverheadRemover" -f "$file" -t "$test_func"
         packageName=$(basename "$package_path")
         fileName=$(basename "$file")
         mkdir -p "advocateResults/$packageName/$fileName/$test_func"
